requirements.txt
Flask==3.0.3
Werkzeug==3.0.4
SQLAlchemy==2.0.49
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.10
Flask-JWT-Extended==4.6.0
bcrypt==4.2.0
python-dotenv==1.0.1
marshmallow==3.23.1
python-binance==1.0.22
gunicorn==23.0.0
